+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * nostr: add `a` tag to zap receipts ([benthecarman])
 * nostr: change NIP-07 `Error::Wasm` variant value from `JsValue` to `String` ([Yuki Kishimoto])
 * nostr: update `EventBuilder::live_event_msg` fingerprint ([Yuki Kishimoto])
+* nostr: set `kind` arg in `EventBuilder::reaction_extended` as optional ([Yuki Kishimoto])
 * pool: increase default kind 3 event limit to `840000` bytes and `10000` tags ([Yuki Kishimoto])
 * pool: improve accuracy of latency calculation ([Yuki Kishimoto])
 * pool: refactoring and adj. `relay` internal module ([Yuki Kishimoto])